not clear what is exactly being checked but the magic number alignment checking looks to be incorrect now that we're using uint32_t's everywhere.  getting misalignment on big-endian 64-bit system in shaders regression, fixed by checking our magic type.
svn:revision:49079
svn:branch:trunk
svn:account:brlcad

+++ src/libbu/badmagic.c
@@ -38,7 +38,7 @@ bu_badmagic(const uint32_t *ptr, uint32_t magic, const char *str, const char *fi
 		 str, file, line);
 	bu_bomb(buf);
     }
-    if (UNLIKELY(((size_t)(ptr)) & (sizeof(unsigned long)-1))) {
+    if (UNLIKELY(((size_t)(ptr)) & (sizeof(uint32_t)-1))) {
 	snprintf(buf, MAGICBUFSIZ, "ERROR: %p mis-aligned %s pointer, file %s, line %d\n",
 		 (void *)ptr, str, file, line);
 	bu_bomb(buf);
